pyproject.toml
[tool.poetry]
name = "homawork-24-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "homawork_24"}]

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.26.0"
coverage = "^7.5.4"
drf-yasg = "^1.21.7"
setuptools = "^70.2.0"
django = "4.2"
django-cors-headers = "^4.4.0"
requests = "^2.32.3"
stripe = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
